requirements.txt
streamlit
requests
beautifulsoup4
urllib3
streamlit-oauth
PyJWT
supabase
